Support tests
Turns out that TS is not that good for broswer code.
In order to test things, we need to run the code in nodejs in some from,
using modules and everything, but the module syntax for node is
different than ES2015 syntax, so I can't generate both node and browser
code from the same source.
Here's the github issue:
https://github.com/microsoft/TypeScript/issues/16577

It's open since 20 fricking 17.

So, in order to make tests work:
For tesing, we're using a seperate tsconfig file to compile things to
node
For browser code, we compile as usual, but theres an ugly PS script that
runs afterwards and fix the import statements in the JS files